+++ xformers/csrc/attention/cuda/fmha/autogen/impl/cutlassB_f16_aligned_k128_dropout.cu
@@ -52,7 +52,7 @@ __global__ void __launch_bounds__(
 fmha_cutlassB_f16_aligned_128x128_k128_dropout_sm80(typename AttentionBackwardKernel<cutlass::arch::Sm80, cutlass::half_t, true, true, true, 128, 128, 128>::Params p) {
 #ifdef __CUDA_ARCH__
 #if __CUDA_ARCH__ >= 800
-#if __CUDA_ARCH__ < 900
+#if __CUDA_ARCH__ < 1000
   if (!p.advance_to_block()) {
     return;
   }
@@ -61,7 +61,7 @@ fmha_cutlassB_f16_aligned_128x128_k128_dropout_sm80(typename AttentionBackwardKe
 #endif
 #endif
     printf(
-        "FATAL: kernel `fmha_cutlassB_f16_aligned_128x128_k128_dropout_sm80` is for sm80-sm90, but was built for sm%d\n",
+        "FATAL: kernel `fmha_cutlassB_f16_aligned_128x128_k128_dropout_sm80` is for sm80-sm100, but was built for sm%d\n",
         int(__CUDA_ARCH__ + 0) / 10);
 #endif
 }
@@ -128,7 +128,7 @@ __global__ void __launch_bounds__(
 fmha_cutlassB_f16_aligned_64x64_k128_dropout_sm80(typename AttentionBackwardKernel<cutlass::arch::Sm80, cutlass::half_t, true, true, false, 64, 64, 128>::Params p) {
 #ifdef __CUDA_ARCH__
 #if __CUDA_ARCH__ >= 800
-#if __CUDA_ARCH__ < 900
+#if __CUDA_ARCH__ < 1000
   if (!p.advance_to_block()) {
     return;
   }
@@ -137,7 +137,7 @@ fmha_cutlassB_f16_aligned_64x64_k128_dropout_sm80(typename AttentionBackwardKern
 #endif
 #endif
     printf(
-        "FATAL: kernel `fmha_cutlassB_f16_aligned_64x64_k128_dropout_sm80` is for sm80-sm90, but was built for sm%d\n",
+        "FATAL: kernel `fmha_cutlassB_f16_aligned_64x64_k128_dropout_sm80` is for sm80-sm100, but was built for sm%d\n",
         int(__CUDA_ARCH__ + 0) / 10);
 #endif
 }