+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fMHA/cutlass (backward): Massive performance improvements when `batch_size * num_heads` is low (10x+)
 - fMHA/cutlass: Further performance improvements for both the forward & backward kernels
 - fMHA (backward): Now dispatching to cutlass when `embed_dim>64`
+- fMHA: Updated Flash-Attention to `v1.0.5`
 ### Fixed
 ### Added
+- fMHA now runs on H100 (support is experimental)
 
 ## [0.0.19] - 2023-04-28
 ### Added